backend/requirements.txt
Django==4.2.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
drf-yasg==1.21.7
requests==2.31.0
python-dotenv==1.0.0